Implement __eq__ for Episode, Season, and Series
Should hopefully fix a bug where Manager.is_episode_queued_for_retrieval(ep)
returns False but ep shows in the download list.  I believe this happens
when an episode is queued for download and the database is subsequently
modified in a way that invalidates the cache, which causes new Episode
objects to be created.

Since this is a blind fix, temporarily log a warning when comparing two equal
objects that don't have the same identiy in order to help troubleshoot.